+++ sdk/azcore/CHANGELOG.md
@@ -1,10 +1,15 @@
 # Release History
 
-## 1.15.0 (2024-10-03)
+## 1.15.0 (2024-10-04)
 
 ### Features Added
 
 * `BearerTokenPolicy` handles CAE claims challenges
+
+### Bugs Fixed
+
+* Omit the `ResponseError.RawResponse` field from JSON marshaling so instances can be marshaled.
+* Fixed an integer overflow in the retry policy.
 
 ## 1.14.0 (2024-08-07)